refactor: task update 웹소켓 enum으로 분리
task 수정은 diff
task 진행 상태 수정은 status
task의 step 변경은 step
taskStepUpdate는 task로 이동
