+++ src/modules/steps/services/steps.service.ts
@@ -14,7 +14,9 @@ import { TaskRepository } from 'src/modules/tasks/repositories/task.repository';
 import { EventBusService } from 'src/infra/event-bus/event-bus.service';
 import { RealTimeEntity, RealTimeType } from 'src/common/response/real-time-response.dto';
 import { EventPayloadDto } from 'src/common/dtos/event-payload.dto';
-import { UpdatedTaskStepDTO, DeletedStepDTO, UpdatedStepDTO } from '../dtos/step-payload.dto';
+import { DeletedStepDTO, UpdatedStepDTO } from '../dtos/step-payload.dto';
+import { UpdatedTaskStepDTO } from 'src/modules/tasks/dtos/task-payload.dto';
+import { TaskUpdatedSubType } from 'src/modules/tasks/enums/task-update-type.enum';
 @Injectable()
 export class StepsService {
     constructor(
@@ -75,7 +77,7 @@ export class StepsService {
         });
         await this.taskRepository.saveWithQueryRunner(qr.manager, partial);
         await this.eventBus.publishAsync(
-            `${RealTimeEntity.TASK}.${RealTimeType.UPDATED}`,
+            `${RealTimeEntity.TASK}.${RealTimeType.UPDATED}.${TaskUpdatedSubType.STEP}`,
             EventPayloadDto.from(RealTimeType.UPDATED, {
                 task: UpdatedTaskStepDTO.from(newStep.project.id, partial.id, newStepId),
             })
