+++ src/modules/tasks/enums/task-update-type.enum.ts
@@ -1,0 +1,5 @@
+export enum TaskUpdatedSubType {
+  DIFF = 'diff',
+  STATUS = 'status',
+  STEP = 'step',
+}
